controller

+++ src/main/java/com/example/berzkoder02/services/ProductService.java
@@ -1,6 +1,7 @@
 package com.example.berzkoder02.services;
 
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.berzkoder02.models.entities.Product;
@@ -14,24 +15,27 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
-    public Product create(Product product) {
+    public Product save(Product product) {
         return productRepository.save(product);
     }
 
     public Product findById(Long id) {
-        return productRepository.findById(id).get();
+        Optional<Product> tmp = productRepository.findById(id);
+        if (!tmp.isPresent()) {
+            return null;
+        }
+        return tmp.get();
     }
 
-    public Iterable<Product> findAll() {
+    public List<Product> findAll() {
         return productRepository.findAll();
     }
 
     public void delete(Long id) {
-        productRepository.deleteById(id);
-    }
-
-    public Product update(Product product) {
-        return productRepository.save(product);
+        Optional<Product> tmp = productRepository.findById(id);
+        if (tmp.isPresent()) {
+            productRepository.deleteById(id);
+        }
     }
 
     public List<Product> findByNameContains(String name) {
